sql 변경

+++ boot/src/main/java/kr/kh/boot/dao/ApiDAO.java
@@ -1,5 +1,7 @@
 package kr.kh.boot.dao;
 
+import java.math.BigDecimal;
+
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -9,7 +11,7 @@ public interface ApiDAO {
 		int countDataByNameAndDate(@Param("name") String name, @Param("date") String date);
 
     public void insertApiData(@Param("name") String name, 
-                              @Param("value") double value, 
+                              @Param("value") BigDecimal value, 
                               @Param("date") String date,
                               @Param("datetime") String datetime);
 }
